UI: Make ToggleBlockRecipe also more clearly show the enabled options.
